Add files via upload
Directed and Undirected are new but not finished.
Graph now has the recommended Exception thrown and int instead of long.

+++ Graph.java
@@ -1,11 +1,11 @@
+/*
+ * Interface of a general Graph <tt>G<V, E></tt> contains a set <tt>V</tt> of vertices and a set <tt>E</tt>
+ * of edges.  Each edge e=<v, u> in E connects v to u. In this case the vertices are only named by numbers 
+ * and the edges are specified by their start- and endpoints.
+ */
 
 public interface Graph {
-	/*
-	 * Interface of a general Graph <tt>G<V, E></tt> contains a set <tt>V</tt> of vertices and a set <tt>E</tt>
-	 * of edges.  Each edge e=<v, u> in E connects v to u. In this case the vertices are only named by numbers 
-	 * and the edges are specified by their start- and endpoints.
-	 */
-
+	
 	/**
 	 * Adds a new vertex with default name (i.e. number) to this graph.
 	 */
@@ -22,7 +22,7 @@ public interface Graph {
 	 * 
 	 * @throws NullPointerException if the specified name is <code> null</code>.
 	 */
-	boolean addVertex(long vertexName);
+	boolean addVertex(int vertexName);
 	
 	/**
 	 * Returns <tt>true</tt> if this graph contains the specified vertex,
@@ -34,7 +34,7 @@ public interface Graph {
 	 * 
 	 * @throws NullPointerException if the specified name is <code> null</code>.
 	 */
-	boolean containsVertex(long vertexName);
+	boolean containsVertex(int vertexName);
 	
 	/**
 	 * Deletes the specified vertex from this graph including all its touching
@@ -51,14 +51,15 @@ public interface Graph {
 	 * 
 	 * @throws NullPointerException if the specified name is <code> null</code>.
 	 */
-	boolean deleteVertex(long vertexName);
+	boolean deleteVertex(int vertexName);
 	
 	/**
 	 * Returns <tt>true</tt> if both specified vertices are adjacent. More formally, if there exists an
 	 * edge (vertexNameStart,vertexNameEnd) in the graph.
 	 * 
 	 * <p>The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of an edge to be tested.
 	 * @param vertexNameEnd name of the endpoint of an edge to be tested.
@@ -66,18 +67,19 @@ public interface Graph {
 	 * @return <tt>true</tt> if the specified vertices are adjacent.
 	 * 
 	 * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
      * @throws NullPointerException if any of the specified vertices is <code>
      * null</code>.
 	 */
-	boolean isAdjacentTo(long vertexNameStart, long vertexNameEnd);
+	boolean isAdjacentTo(int vertexNameStart, int vertexNameEnd);
 	
 	/** Adds a new edge between the vertices vertexNameStart and vertexNameEnd with default weight 1.
 	 * If this graph already contains an edge between those vertices, the call leaves this graph unchanged and returns <tt>false</tt>.
      * If the edge was added to the graph, returns <code>true</code>.
 	 * 
 	 * <p>The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of the edge to be added.
 	 * @param vertexNameEnd name of the endpoint of the edge to be added.
@@ -86,18 +88,19 @@ public interface Graph {
      * edge.
      *
      * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
      * @throws NullPointerException if any of the specified vertices is <code>
      * null</code>.
 	 */
-	boolean addEdge(long vertexNameStart, long vertexNameEnd);
+	boolean addEdge(int vertexNameStart, int vertexNameEnd);
 	
 	/** Adds a new edge between the vertices vertexNameStart and vertexNameEnd with specified weight.
 	 * If this graph already contains an edge between those vertices, the call leaves this graph unchanged and returns <tt>false</tt>.
      * If the edge was added to the graph, returns <code>true</code>.
 	 * 
 	 * <p>The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of the edge to be added.
 	 * @param vertexNameEnd name of the endpoint of the edge to be added.
@@ -107,17 +110,18 @@ public interface Graph {
      * edge.
      *
      * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
      * @throws NullPointerException if any of the specified vertices is <code>
      * null</code>.
 	 */
-	boolean addEdge(long vertexNameStart, long vertexNameEnd, int edgeWeight);
+	boolean addEdge(int vertexNameStart, int vertexNameEnd, int edgeWeight);
 	
 	/**
 	 * Returns <tt>true</tt> if this graph contains the specified edge between the vertices vertexNameStart and vertexNameEnd.
 	 * 
 	 * <p>The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of the edge to be tested.
 	 * @param vertexNameEnd name of the endpoint of the edge to be tested.
@@ -125,10 +129,10 @@ public interface Graph {
 	 * @return <tt>true</tt> if this graph contains the specified edge.
 	 *
 	 * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
 	 * @throws NullPointerException if any of the specified vertices is <code> null</code>.
 	 */
-	boolean containsEdge(long vertexNameStart, long vertexNameEnd);
+	boolean containsEdge(int vertexNameStart, int vertexNameEnd);
 	
 	/**
 	 * Deletes the edge specified by their start- and endpoint from this graph. All incident vertices
@@ -138,7 +142,8 @@ public interface Graph {
      * (The graph will not contain the specified edge once the call returns).
 	 * 
 	 * <p>The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of the edge to be deleted.
 	 * @param vertexNameEnd name of the endpoint of the edge to be deleted.
@@ -147,17 +152,18 @@ public interface Graph {
      * <code>false</code> otherwise.
      * 
      * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
 	 * @throws NullPointerException if any of the specified vertices is <code> null</code>.
 	 */
-	boolean deleteEdge(long vertexNameStart, long vertexNameEnd);
+	boolean deleteEdge(int vertexNameStart, int vertexNameEnd);
 	
 	/**
 	 * Returns the weight of the edge specified by their start- and endpoint.
 	 *
 	 * <p> If the edge does not exist NoSuchElementException is thrown.
 	 * The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of the edge whose weight shall be returned.
 	 * @param vertexNameEnd name of the endpoint of the edge whose weight shall be returned.
@@ -166,10 +172,10 @@ public interface Graph {
 	 * 
 	 * @throws NoSuchElementException if the specified edge is not found in the graph.
 	 * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
 	 * @throws NullPointerException if any of the specified vertices is <code> null</code>.
 	 */
-	int getEdgeWeight(long vertexNameStart, long vertexNameEnd);
+	int getEdgeWeight(int vertexNameStart, int vertexNameEnd);
 	
 	/**
 	
@@ -177,7 +183,8 @@ public interface Graph {
 	 * 
 	 * <p>If the edge does not exist NoSuchElementException is thrown.
 	 * The start and end vertices must already be contained in this graph.
-	 * If they are not found in graph IllegalArgumentException is thrown.</p>
+	 * If they are not found in graph IllegalArgumentException is thrown.
+	 * If they are the same IllegalArgumentException is thrown as well.</p>
 	 * 
 	 * @param vertexNameStart name of the startpoint of the edge whose weight shall be changed.
 	 * @param vertexNameEnd name of the endpoint of the edge whose weight shall be changed.
@@ -185,10 +192,9 @@ public interface Graph {
 	 * 
 	 * @throws NoSuchElementException if the specified edge is not found in the graph.
 	 * @throws IllegalArgumentException if start or end vertices are not
-     * found in the graph.
+     * found in the graph or if start- and endpoint are the same.
 	 * @throws NullPointerException if any of the specified vertices or the new weight is <code> null</code>.
 	 */
-	void setEdgeWeight(long vertexNameStart, long vertexNameEnd, int newEdgeWeight);
-	
+	void setEdgeWeight(int vertexNameStart, int vertexNameEnd, int newEdgeWeight);
 	
 }
